Quest data is sent by the server QUEST_LOG_UPDATE UI event is sent to the UI on quest log changes from the server Fixed a potential UI crash on the client More new mirror handler functions in world state for field changes like attribute points

+++ src/shared/game/quest_info.cpp
@@ -10,8 +10,8 @@ namespace mmo
 	{
 		writer << io::write<uint64>(itemInfo.id)
 			<< io::write_dynamic_range<uint8>(itemInfo.title)
-			<< io::write_dynamic_range<uint8>(itemInfo.summary)
-			<< io::write_dynamic_range<uint8>(itemInfo.description);
+			<< io::write_dynamic_range<uint16>(itemInfo.summary)
+			<< io::write_dynamic_range<uint16>(itemInfo.description);
 
 		return writer;
 	}
@@ -20,8 +20,8 @@ namespace mmo
 	{
 		reader >> io::read<uint64>(outItemInfo.id)
 			>> io::read_container<uint8>(outItemInfo.title)
-			>> io::read_container<uint8>(outItemInfo.summary)
-			>> io::read_container<uint8>(outItemInfo.description);
+			>> io::read_container<uint16>(outItemInfo.summary)
+			>> io::read_container<uint16>(outItemInfo.description);
 
 		return reader;
 	}
